Version_4
After spending some time trying to see why the music activity was so laggy anyways and i guess it was just the seekbar had a wrong style and that was the cause.

Also made it so that you had 100% to allow the app to see your external files because well it wont even work without it so why bother.